pyproject.toml
[project]
name = "orquestra"
version = "0.1.0"
description = "Orchestrate agents and tools in perfect harmony - A modern AI agent framework with multi-provider support"
readme = "README.md"
authors = [
    { name = "Marcos Oliveira", email = "marcosf63@gmail.com" }
]
requires-python = ">=3.13"
keywords = ["ai", "agents", "llm", "framework", "orchestration", "multi-provider"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.40.0",
]
gemini = [
    "google-generativeai>=0.8.0",
]
ollama = [
    "ollama>=0.3.0",
]
all = [
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "ollama>=0.3.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "black>=24.0.0",
]
search = [
    "duckduckgo-search>=8.1.1",
]

[project.urls]
Homepage = "https://github.com/marcosf63/orquestra"
Documentation = "https://github.com/marcosf63/orquestra#readme"
Repository = "https://github.com/marcosf63/orquestra"
Issues = "https://github.com/marcosf63/orquestra/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.0",
]
